Refactor `endpointish` into a subpackage
There are three distinct "pillars" of this functionality:
- error handling types
- EndpointType
- the main Endpointish class

EndpointType has to be moved from `globus_cli.services.transfer` for
this to make sense, but it turns out to be fully encapsulated within
Endpointish now.
