+++ src/globus_cli/endpointish/endpoint_type.py
